refactor deduplicate page
