--- src/types/deduplicating/index.ts
@@ -1,6 +1,0 @@
-export interface DeduplicateContextType {
-    clubSameTimeFilesOnly: boolean;
-    setClubSameTimeFilesOnly: (clubSameTimeFilesOnly: boolean) => void;
-    fileSizeMap: Map<number, number>;
-    state: boolean;
-}
